requirements.txt
PySide2==5.13.2
shiboken2==5.13.2
Pillow==9.2.0
pytest==7.1.2